requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
dj-database-url==0.5.0
Django==2.1.7
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-heroku==0.3.1
django-nose==1.4.6
djangorestframework==3.9.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
nose==1.3.7
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
readtime==1.1.1
python-dotenv==0.1.0
